+++ ChangeLog.md
@@ -5,6 +5,8 @@
 - Fix a bug that caused another code-leftover to be displayed in the finalization proof
   line of the gas rewards section of the `consensus show-chain-parameters` output. This
   is only relevant when the queried chain runs protocol version 6.
+- Add `raw GetBlockTransactionEvents` that prints the list of transaction outcomes
+  in a given block.
 
 ## 5.1.1